+++ src/plone/api/content.py
@@ -110,7 +110,10 @@ def create(
     if IBaseObject.providedBy(content):
         # Will finish Archetypes content item creation process,
         # rename-after-creation and such
-        content.processForm(values=kwargs)
+        # Passing values as a dict with None values so values set by
+        # invokeFactory don't get overridden.
+        # None: None is required so that bool(values) is True.
+        content.processForm(values={None: None})
 
     if not id or (safe_id and id):
         # Create a new id from title